Split methods checking item availability into a seperate service along with their tests
