+++ ausleih-service/src/main/java/hhu/ausleihservice/web/ItemService.java
@@ -17,9 +17,11 @@ import java.util.stream.Stream;
 public class ItemService {
 
 	private ItemRepository items;
+	private ItemAvailabilityService itemAvailabilityService;
 
-	ItemService(ItemRepository itemRep) {
+	ItemService(ItemRepository itemRep, ItemAvailabilityService itemAvailabilityService) {
 		this.items = itemRep;
+		this.itemAvailabilityService = itemAvailabilityService;
 	}
 
 	Item findByID(long id) {
@@ -33,45 +35,6 @@ public class ItemService {
 
 	List<Item> findAll() {
 		return items.findAll();
-	}
-
-	private boolean isInPeriod(LocalDate date, LocalDate start, LocalDate end) {
-		return (!date.isBefore(start) && !date.isAfter(end));
-	}
-
-	public boolean isAvailable(Item item) {
-		return isAvailable(item, LocalDate.now());
-	}
-
-	boolean isAvailable(Item item, LocalDate date) {
-		LocalDate availableFrom = item.getAvailableFrom();
-		LocalDate availableTill = item.getAvailableTill();
-		Set<Ausleihe> ausleihen = item.getAusleihen();
-
-		if (!isInPeriod(date, availableFrom, availableTill)) {
-			return false;
-		}
-		for (Ausleihe ausleihe : ausleihen) {
-			LocalDate startDatum = ausleihe.getStartDatum();
-			LocalDate endDatum = ausleihe.getEndDatum();
-			if (isInPeriod(date, startDatum, endDatum)) {
-				return false;
-			}
-		}
-		return true;
-	}
-
-	//Format of input is "YYYY-MM-DD"
-	public boolean isAvailableFromTill(Item item, String from, String till) {
-		LocalDate temp = LocalDate.parse(from);
-		LocalDate end = LocalDate.parse(till);
-		while (!temp.equals(end.plusDays(1))) {
-			if (!isAvailable(item, temp)) {
-				return false;
-			}
-			temp = temp.plusDays(1);
-		}
-		return true;
 	}
 
 	private boolean containsArray(String string, String[] array) {
@@ -127,7 +90,7 @@ public class ItemService {
 		listStream = listStream.filter(item -> item.getKautionswert() <= kautionswertMax);
 
 		listStream = listStream.filter(
-				item -> isAvailableFromTill(item, availableMin, availableMax)
+				item -> itemAvailabilityService.isAvailableFromTill(item, availableMin, availableMax)
 		);
 
 		List<Item> list = listStream.collect(Collectors.toList());
